Remove superseded source parser
- Route commands through the current row parser

- Move AST cleanup into its own native unit

- Update native smoke and metrics scripts
