+++ native/zero-c/src/fs.c
@@ -191,11 +191,6 @@ bool z_write_binary_file(const char *path, const unsigned char *data, size_t len
   return true;
 }
 
-static bool is_directory(const char *path) {
-  struct stat st;
-  return stat(path, &st) == 0 && S_ISDIR(st.st_mode);
-}
-
 static bool ends_with(const char *text, const char *suffix) {
   size_t text_len = strlen(text);
   size_t suffix_len = strlen(suffix);
@@ -279,55 +274,6 @@ static unsigned long long fs_mix_hash_text(unsigned long long hash, const char *
   return hash;
 }
 
-static void source_input_push_file(SourceInput *input, const char *path) {
-  input->source_files = z_checked_reallocarray(input->source_files, input->source_file_count + 1, sizeof(char *));
-  input->source_files[input->source_file_count++] = z_strdup(path);
-}
-
-static void source_input_push_source_line(SourceInput *input, const char *path, int original_line) {
-  input->source_line_paths = z_checked_reallocarray(input->source_line_paths, input->source_line_count + 1, sizeof(char *));
-  input->source_line_numbers = z_checked_reallocarray(input->source_line_numbers, input->source_line_count + 1, sizeof(int));
-  input->source_line_paths[input->source_line_count] = z_strdup(path);
-  input->source_line_numbers[input->source_line_count] = original_line > 0 ? original_line : 1;
-  input->source_line_count++;
-}
-
-static void source_input_push_final_source_line(SourceInput *input) {
-  if (!input || input->source_line_count == 0) return;
-  source_input_push_source_line(input, input->source_line_paths[input->source_line_count - 1], input->source_line_numbers[input->source_line_count - 1]);
-}
-
-static void source_input_push_import(SourceInput *input, const char *name) {
-  input->imports = z_checked_reallocarray(input->imports, input->import_count + 1, sizeof(char *));
-  input->imports[input->import_count++] = z_strdup(name);
-}
-
-static void source_input_push_module(SourceInput *input, const char *name, const char *path) {
-  input->module_names = z_checked_reallocarray(input->module_names, input->module_count + 1, sizeof(char *));
-  input->module_paths = z_checked_reallocarray(input->module_paths, input->module_count + 1, sizeof(char *));
-  input->module_names[input->module_count] = z_strdup(name);
-  input->module_paths[input->module_count] = z_strdup(path);
-  input->module_count++;
-}
-
-static void source_input_push_import_edge(SourceInput *input, const char *from, const char *to, const char *path, const char *source_path, int line, int column, int length) {
-  input->import_from = z_checked_reallocarray(input->import_from, input->import_edge_count + 1, sizeof(char *));
-  input->import_to = z_checked_reallocarray(input->import_to, input->import_edge_count + 1, sizeof(char *));
-  input->import_paths = z_checked_reallocarray(input->import_paths, input->import_edge_count + 1, sizeof(char *));
-  input->import_source_paths = z_checked_reallocarray(input->import_source_paths, input->import_edge_count + 1, sizeof(char *));
-  input->import_lines = z_checked_reallocarray(input->import_lines, input->import_edge_count + 1, sizeof(int));
-  input->import_columns = z_checked_reallocarray(input->import_columns, input->import_edge_count + 1, sizeof(int));
-  input->import_lengths = z_checked_reallocarray(input->import_lengths, input->import_edge_count + 1, sizeof(int));
-  input->import_from[input->import_edge_count] = z_strdup(from);
-  input->import_to[input->import_edge_count] = z_strdup(to);
-  input->import_paths[input->import_edge_count] = z_strdup(path ? path : "");
-  input->import_source_paths[input->import_edge_count] = z_strdup(source_path ? source_path : "");
-  input->import_lines[input->import_edge_count] = line > 0 ? line : 1;
-  input->import_columns[input->import_edge_count] = column > 0 ? column : 1;
-  input->import_lengths[input->import_edge_count] = length > 0 ? length : 1;
-  input->import_edge_count++;
-}
-
 static void source_input_push_dependency(SourceInput *input, const ZManifestDependency *dep, const char *resolved_manifest, const char *resolved_name, const char *resolved_version, const char *status, bool direct) {
   input->dependencies = z_checked_reallocarray(input->dependencies, input->dependency_count + 1, sizeof(SourceDependency));
   SourceDependency *item = &input->dependencies[input->dependency_count++];
@@ -361,419 +307,6 @@ static const char *source_input_package_version(const SourceInput *input, const 
     }
   }
   return NULL;
-}
-
-static bool source_input_push_symbol(SourceInput *input, const char *module, const char *kind, const char *name, bool is_public, ZDiag *diag) {
-  if (is_public) {
-    for (size_t i = 0; i < input->symbol_count; i++) {
-      if (input->symbol_public[i] && strcmp(input->symbol_names[i], name) == 0 && strcmp(input->symbol_modules[i], module) != 0) {
-        diag->code = 7003;
-        diag->path = input->source_file;
-        diag->line = 1;
-        diag->column = 1;
-        snprintf(diag->message, sizeof(diag->message), "duplicate public symbol '%s'", name);
-        snprintf(diag->expected, sizeof(diag->expected), "unique public symbol names across imported modules");
-        snprintf(diag->actual, sizeof(diag->actual), "%s also exported by %s", name, input->symbol_modules[i]);
-        snprintf(diag->help, sizeof(diag->help), "rename one symbol or keep it private inside its module");
-        return false;
-      }
-    }
-  }
-  input->symbol_names = z_checked_reallocarray(input->symbol_names, input->symbol_count + 1, sizeof(char *));
-  input->symbol_modules = z_checked_reallocarray(input->symbol_modules, input->symbol_count + 1, sizeof(char *));
-  input->symbol_kinds = z_checked_reallocarray(input->symbol_kinds, input->symbol_count + 1, sizeof(char *));
-  input->symbol_public = z_checked_reallocarray(input->symbol_public, input->symbol_count + 1, sizeof(bool));
-  input->symbol_names[input->symbol_count] = z_strdup(name);
-  input->symbol_modules[input->symbol_count] = z_strdup(module);
-  input->symbol_kinds[input->symbol_count] = z_strdup(kind ? kind : "unknown");
-  input->symbol_public[input->symbol_count] = is_public;
-  input->symbol_count++;
-  return true;
-}
-
-static char *module_name_from_path(const char *src_root, const char *path) {
-  const char *relative = path;
-  size_t root_len = src_root ? strlen(src_root) : 0;
-  if (src_root && strncmp(path, src_root, root_len) == 0) {
-    relative = path + root_len;
-    if (*relative == '/') relative++;
-  } else {
-    const char *slash = strrchr(path, '/');
-    relative = slash ? slash + 1 : path;
-  }
-  size_t len = strlen(relative);
-  if (len > 2 && strcmp(relative + len - 2, ".0") == 0) len -= 2;
-  if (len >= 4 && strcmp(relative + len - 4, "/mod") == 0) len -= 4;
-  ZBuf buf;
-  zbuf_init(&buf);
-  for (size_t i = 0; i < len; i++) zbuf_append_char(&buf, relative[i] == '/' ? '.' : relative[i]);
-  if (buf.len == 0) zbuf_append(&buf, "main");
-  return buf.data;
-}
-
-static bool source_input_has_file(SourceInput *input, const char *path) {
-  for (size_t i = 0; i < input->source_file_count; i++) {
-    if (strcmp(input->source_files[i], path) == 0) return true;
-  }
-  return false;
-}
-
-static bool import_stack_contains(char **stack, size_t len, const char *path) {
-  for (size_t i = 0; i < len; i++) {
-    if (strcmp(stack[i], path) == 0) return true;
-  }
-  return false;
-}
-
-static char *module_path_to_source(const char *src_root, const char *module_name) {
-  ZBuf relative;
-  zbuf_init(&relative);
-  for (const char *cursor = module_name; *cursor; cursor++) {
-    zbuf_append_char(&relative, *cursor == '.' ? '/' : *cursor);
-  }
-  zbuf_append(&relative, ".0");
-  char *file_path = join_path(src_root, relative.data);
-  if (file_exists(file_path)) {
-    zbuf_free(&relative);
-    return file_path;
-  }
-  free(file_path);
-  if (relative.len >= 2 && strcmp(relative.data + relative.len - 2, ".0") == 0) {
-    relative.data[relative.len - 2] = 0;
-    relative.len -= 2;
-  }
-  char *dir_path = join_path(src_root, relative.data);
-  char *mod_path = join_path(dir_path, "mod.0");
-  free(dir_path);
-  zbuf_free(&relative);
-  if (file_exists(mod_path)) return mod_path;
-  free(mod_path);
-  return NULL;
-}
-
-static void append_source_without_imports(SourceInput *input, ZBuf *buf, const char *path, const char *source) {
-  zbuf_append(buf, "// file: ");
-  zbuf_append(buf, path);
-  zbuf_append(buf, "\n");
-  source_input_push_source_line(input, path, 1);
-  const char *line = source;
-  int original_line = 1;
-  while (*line) {
-    const char *end = strchr(line, '\n');
-    size_t len = end ? (size_t)(end - line) : strlen(line);
-    const char *start = line;
-    while (len > 0 && isspace((unsigned char)*start)) {
-      start++;
-      len--;
-    }
-    bool is_legacy_import = len >= 7 && strncmp(start, "import ", 7) == 0;
-    if (!is_legacy_import) {
-      zbuf_appendf(buf, "%.*s\n", (int)(end ? (size_t)(end - line) : strlen(line)), line);
-      source_input_push_source_line(input, path, original_line);
-    }
-    original_line++;
-    if (!end) break;
-    line = end + 1;
-  }
-  zbuf_append(buf, "\n");
-  source_input_push_source_line(input, path, original_line);
-}
-
-static bool resolve_imported_source(const char *path, const char *src_root, SourceInput *input, ZBuf *combined, ZDiag *diag, char ***stack, size_t *stack_len);
-
-static bool scan_top_level_symbols(const char *source, const char *module_name, SourceInput *input, ZDiag *diag) {
-  const char *line = source;
-  while (*line && diag->code == 0) {
-    const char *end = strchr(line, '\n');
-    size_t len = end ? (size_t)(end - line) : strlen(line);
-    const char *start = line;
-    while (len > 0 && isspace((unsigned char)*start)) {
-      start++;
-      len--;
-    }
-    bool is_public = false;
-    if (len > 4 && strncmp(start, "pub ", 4) == 0) {
-      is_public = true;
-      start += 4;
-      len -= 4;
-      while (len > 0 && isspace((unsigned char)*start)) {
-        start++;
-        len--;
-      }
-    }
-    if (len > 7 && strncmp(start, "extern ", 7) == 0) {
-      start += 7;
-      len -= 7;
-      while (len > 0 && isspace((unsigned char)*start)) {
-        start++;
-        len--;
-      }
-    } else if (len > 7 && strncmp(start, "packed ", 7) == 0) {
-      start += 7;
-      len -= 7;
-      while (len > 0 && isspace((unsigned char)*start)) {
-        start++;
-        len--;
-      }
-    }
-    const char *after_keyword = NULL;
-    const char *kind = NULL;
-    const char *keywords[] = {"fun ", "shape ", "interface ", "enum ", "choice ", "const ", "type ", NULL};
-    const char *kinds[] = {"function", "shape", "interface", "enum", "choice", "const", "alias", NULL};
-    for (int i = 0; keywords[i]; i++) {
-      size_t keyword_len = strlen(keywords[i]);
-      if (len > keyword_len && strncmp(start, keywords[i], keyword_len) == 0) {
-        after_keyword = start + keyword_len;
-        kind = kinds[i];
-        break;
-      }
-    }
-    if (after_keyword) {
-      size_t name_len = 0;
-      while (after_keyword[name_len] && (isalnum((unsigned char)after_keyword[name_len]) || after_keyword[name_len] == '_')) name_len++;
-      if (name_len > 0) {
-        char *name = z_strndup(after_keyword, name_len);
-        bool ok = source_input_push_symbol(input, module_name, kind, name, is_public, diag);
-        free(name);
-        if (!ok) return false;
-      }
-    }
-    if (!end) break;
-    line = end + 1;
-  }
-  return diag->code == 0;
-}
-
-static void format_cycle_chain(const char *src_root, char **stack, size_t stack_len, const char *path, ZBuf *out) {
-  bool started = false;
-  for (size_t i = 0; i < stack_len; i++) {
-    if (!started && strcmp(stack[i], path) != 0) continue;
-    started = true;
-    if (out->len > 0) zbuf_append(out, " -> ");
-    char *module = module_name_from_path(src_root, stack[i]);
-    zbuf_append(out, module);
-    free(module);
-  }
-  if (out->len > 0) zbuf_append(out, " -> ");
-  char *module = module_name_from_path(src_root, path);
-  zbuf_append(out, module);
-  free(module);
-}
-
-static bool import_ident_start(char ch) {
-  return isalpha((unsigned char)ch) || ch == '_';
-}
-
-static bool import_ident_continue(char ch) {
-  return isalnum((unsigned char)ch) || ch == '_';
-}
-
-static bool import_ident_is_keyword(const char *text, size_t len) {
-  const char *keywords[] = {
-    "as", "break", "check", "choice", "const", "continue", "defer", "else", "enum", "export", "extern", "false", "for", "fun",
-    "if", "import", "in", "let", "match", "meta", "mut", "null", "packed", "pub",
-    "raise", "raises", "rescue", "return", "shape", "static", "test", "true", "type",
-    "use", "var", "while", NULL
-  };
-  for (int i = 0; keywords[i]; i++) {
-    if (strlen(keywords[i]) == len && strncmp(text, keywords[i], len) == 0) return true;
-  }
-  return false;
-}
-
-static bool parse_use_import_ident_segment(const char *text, size_t len, size_t *pos, const char **segment_out, size_t *segment_len_out) {
-  if (*pos >= len || !import_ident_start(text[*pos])) return false;
-  size_t start = *pos;
-  (*pos)++;
-  while (*pos < len && import_ident_continue(text[*pos])) (*pos)++;
-  if (import_ident_is_keyword(text + start, *pos - start)) return false;
-  if (segment_out) *segment_out = text + start;
-  if (segment_len_out) *segment_len_out = *pos - start;
-  return true;
-}
-
-static bool import_line_comment_at(const char *text, size_t pos, size_t len) {
-  return pos + 1 < len && text[pos] == '/' && text[pos + 1] == '/';
-}
-
-static void import_line_skip_ws(const char *text, size_t len, size_t *pos) {
-  while (*pos < len && isspace((unsigned char)text[*pos])) (*pos)++;
-}
-
-static void import_line_append_span(ZBuf *buf, const char *text, size_t len) {
-  for (size_t i = 0; i < len; i++) zbuf_append_char(buf, text[i]);
-}
-
-static bool parse_use_import_line_module(const char *text, size_t len, char **module_out, int *length_out) {
-  size_t pos = 0;
-  import_line_skip_ws(text, len, &pos);
-  ZBuf module;
-  zbuf_init(&module);
-  bool wrote_segment = false;
-  size_t last_token_end = 0;
-  for (;;) {
-    const char *segment = NULL;
-    size_t segment_len = 0;
-    if (!parse_use_import_ident_segment(text, len, &pos, &segment, &segment_len)) {
-      zbuf_free(&module);
-      return false;
-    }
-    if (wrote_segment) zbuf_append_char(&module, '.');
-    import_line_append_span(&module, segment, segment_len);
-    wrote_segment = true;
-    last_token_end = pos;
-    import_line_skip_ws(text, len, &pos);
-    if (pos < len && text[pos] == '.') {
-      pos++;
-      import_line_skip_ws(text, len, &pos);
-      continue;
-    }
-    break;
-  }
-  import_line_skip_ws(text, len, &pos);
-  if (import_line_comment_at(text, pos, len)) pos = len;
-  if (pos < len) {
-    if (pos + 2 > len || strncmp(text + pos, "as", 2) != 0) {
-      zbuf_free(&module);
-      return false;
-    }
-    pos += 2;
-    if (pos >= len || !isspace((unsigned char)text[pos])) {
-      zbuf_free(&module);
-      return false;
-    }
-    import_line_skip_ws(text, len, &pos);
-    if (!parse_use_import_ident_segment(text, len, &pos, NULL, NULL)) {
-      zbuf_free(&module);
-      return false;
-    }
-    last_token_end = pos;
-    import_line_skip_ws(text, len, &pos);
-    if (import_line_comment_at(text, pos, len)) pos = len;
-    if (pos < len) {
-      zbuf_free(&module);
-      return false;
-    }
-  }
-  *module_out = module.data ? module.data : z_strdup("");
-  if (length_out) *length_out = 3 + (int)last_token_end;
-  return true;
-}
-
-static void set_import_source_diag(ZDiag *diag, int code, const char *path, int line, int column, int length) {
-  diag->code = code;
-  diag->path = z_strdup(path ? path : "");
-  diag->line = line > 0 ? line : 1;
-  diag->column = column > 0 ? column : 1;
-  diag->length = length > 0 ? length : 1;
-}
-
-static bool scan_imports_and_append_dependencies(const char *source, const char *src_root, const char *current_module, const char *current_path, SourceInput *input, ZBuf *combined, ZDiag *diag, char ***stack, size_t *stack_len) {
-  const char *line = source;
-  int original_line = 1;
-  while (*line && diag->code == 0) {
-    const char *end = strchr(line, '\n');
-    size_t len = end ? (size_t)(end - line) : strlen(line);
-    const char *start = line;
-    while (len > 0 && isspace((unsigned char)*start)) {
-      start++;
-      len--;
-    }
-    char *module_name = NULL;
-    int import_column = (int)(start - line) + 1;
-    int import_length = (int)len;
-    if (len > 3 && strncmp(start, "use", 3) == 0 && isspace((unsigned char)start[3])) {
-      parse_use_import_line_module(start + 3, len - 3, &module_name, &import_length);
-    } else if (len >= 7 && strncmp(start, "import ", 7) == 0) {
-      const char *module = start + 7;
-      size_t module_len = len - 7;
-      while (module_len > 0 && isspace((unsigned char)module[module_len - 1])) module_len--;
-      const char *as_kw = NULL;
-      for (size_t i = 0; i + 4 <= module_len; i++) {
-        if (strncmp(module + i, " as ", 4) == 0) {
-          as_kw = module + i;
-          break;
-        }
-      }
-      if (as_kw) module_len = (size_t)(as_kw - module);
-      module_name = z_strndup(module, module_len);
-      import_length = 7 + (int)module_len;
-    }
-    if (module_name) {
-      if (strncmp(module_name, "std.", 4) == 0) {
-        free(module_name);
-      } else {
-        source_input_push_import(input, module_name);
-        char *module_path = module_path_to_source(src_root, module_name);
-        if (!module_path) {
-          set_import_source_diag(diag, 7001, current_path, original_line, import_column, import_length);
-          snprintf(diag->message, sizeof(diag->message), "unknown package-local import '%s'", module_name);
-          snprintf(diag->expected, sizeof(diag->expected), "src/%s.0 or src/%s/mod.0", module_name, module_name);
-          snprintf(diag->actual, sizeof(diag->actual), "missing source file");
-          snprintf(diag->help, sizeof(diag->help), "create the module source file or remove the import");
-          free(module_name);
-          return false;
-        }
-        if (import_stack_contains(*stack, *stack_len, module_path)) {
-          ZBuf chain;
-          zbuf_init(&chain);
-          format_cycle_chain(src_root, *stack, *stack_len, module_path, &chain);
-          set_import_source_diag(diag, 7002, current_path, original_line, import_column, import_length);
-          snprintf(diag->message, sizeof(diag->message), "import cycle detected");
-          snprintf(diag->actual, sizeof(diag->actual), "%s", chain.data ? chain.data : module_path);
-          snprintf(diag->help, sizeof(diag->help), "break the module cycle by moving shared declarations into a third module");
-          zbuf_free(&chain);
-          free(module_path);
-          free(module_name);
-          return false;
-        }
-        source_input_push_import_edge(input, current_module, module_name, module_path, current_path, original_line, import_column, import_length);
-        bool ok = resolve_imported_source(module_path, src_root, input, combined, diag, stack, stack_len);
-        free(module_path);
-        free(module_name);
-        if (!ok) return false;
-      }
-    }
-    if (!end) break;
-    line = end + 1;
-    original_line++;
-  }
-  return diag->code == 0;
-}
-
-static bool resolve_imported_source(const char *path, const char *src_root, SourceInput *input, ZBuf *combined, ZDiag *diag, char ***stack, size_t *stack_len) {
-  if (source_input_has_file(input, path)) return true;
-  if (import_stack_contains(*stack, *stack_len, path)) {
-    ZBuf chain;
-    zbuf_init(&chain);
-    format_cycle_chain(src_root, *stack, *stack_len, path, &chain);
-    diag->code = 7002;
-    diag->path = z_strdup(input->source_file);
-    diag->line = 1;
-    diag->column = 1;
-    diag->length = 1;
-    snprintf(diag->message, sizeof(diag->message), "import cycle detected");
-    snprintf(diag->actual, sizeof(diag->actual), "%s", chain.data ? chain.data : path);
-    snprintf(diag->help, sizeof(diag->help), "break the module cycle by moving shared declarations into a third module");
-    zbuf_free(&chain);
-    return false;
-  }
-  char *source = z_read_file(path, diag);
-  if (!source) return false;
-  char *module_name = module_name_from_path(src_root, path);
-  *stack = z_checked_reallocarray(*stack, *stack_len + 1, sizeof(char *));
-  (*stack)[(*stack_len)++] = z_strdup(path);
-  bool ok = scan_imports_and_append_dependencies(source, src_root, module_name, path, input, combined, diag, stack, stack_len);
-  free((*stack)[--(*stack_len)]);
-  if (ok) ok = scan_top_level_symbols(source, module_name, input, diag);
-  if (ok) {
-    source_input_push_file(input, path);
-    source_input_push_module(input, module_name, path);
-    append_source_without_imports(input, combined, path, source);
-  }
-  free(module_name);
-  free(source);
-  return ok;
 }
 
 static const char *json_skip_ws(const char *cursor) {
@@ -1350,80 +883,6 @@ static bool resolve_manifest_dependencies(const char *manifest_path, const ZMani
     if (!ok) return false;
   }
   return true;
-}
-
-bool z_resolve_source(const char *input, SourceInput *out, ZDiag *diag) {
-  char *manifest_path = NULL;
-  if (is_directory(input)) {
-    manifest_path = join_path(input, "zero.json");
-  } else if (ends_with(input, "zero.json")) {
-    manifest_path = z_strdup(input);
-  }
-
-  if (manifest_path) {
-    char *manifest = z_read_file(manifest_path, diag);
-    if (!manifest) {
-      // Keep manifest_path alive for the diagnostic; callers print it after return.
-      return false;
-    }
-    ZManifest parsed_manifest = {0};
-    if (!z_parse_manifest_json(manifest, &parsed_manifest, diag)) {
-      diag->path = manifest_path;
-      free(manifest);
-      // Keep manifest_path alive for the diagnostic; callers print it after return.
-      return false;
-    }
-    if (!z_resolve_package_metadata(manifest_path, manifest, &parsed_manifest, out, diag)) {
-      z_free_manifest(&parsed_manifest);
-      free(manifest);
-      free(manifest_path);
-      return false;
-    }
-    out->source_file = join_path(out->package_root, parsed_manifest.main_path);
-    if (!file_exists(out->source_file)) {
-      diag->code = 2002;
-      diag->path = out->manifest_path ? out->manifest_path : manifest_path;
-      diag->line = 1;
-      diag->column = 1;
-      snprintf(diag->message, sizeof(diag->message), "target main source does not exist");
-      snprintf(diag->expected, sizeof(diag->expected), "%s", out->source_file);
-      snprintf(diag->actual, sizeof(diag->actual), "missing source file");
-      snprintf(diag->help, sizeof(diag->help), "create the main source file or update targets.cli.main");
-      z_free_manifest(&parsed_manifest);
-      free(manifest);
-      free(manifest_path);
-      return false;
-    }
-    ZBuf source;
-    zbuf_init(&source);
-    char *src_dir = join_path(out->package_root, "src");
-    char **stack = NULL;
-    size_t stack_len = 0;
-    resolve_imported_source(out->source_file, src_dir, out, &source, diag, &stack, &stack_len);
-    if (diag->code == 0) source_input_push_final_source_line(out);
-    free(stack);
-    free(src_dir);
-    out->source = diag->code == 0 ? source.data : NULL;
-    if (diag->code != 0) zbuf_free(&source);
-    z_free_manifest(&parsed_manifest);
-    free(manifest);
-    free(manifest_path);
-    return out->source != NULL;
-  }
-
-  out->source_file = z_strdup(input);
-  ZBuf source;
-  zbuf_init(&source);
-  char *dir = dirname_of(input);
-  char **stack = NULL;
-  size_t stack_len = 0;
-  resolve_imported_source(out->source_file, dir, out, &source, diag, &stack, &stack_len);
-  if (diag->code == 0) source_input_push_final_source_line(out);
-  free(stack);
-  free(dir);
-  out->source = diag->code == 0 ? source.data : NULL;
-  if (diag->code != 0) zbuf_free(&source);
-  return out->source != NULL;
 }
 
 bool z_map_source_diag(const SourceInput *input, ZDiag *diag) {
