+++ scripts/compiler-metrics.mts
@@ -4,8 +4,8 @@ const sourceFiles = [
   "native/zero-c/src/checker.c",
   "native/zero-c/src/main.c",
   "native/zero-c/src/ir.c",
-  "native/zero-c/src/parser.c",
-  "native/zero-c/src/lexer.c",
+  "native/zero-c/src/row_syntax.c",
+  "native/zero-c/src/ast.c",
   "native/zero-c/src/emit_macho64.c",
   "native/zero-c/src/emit_elf64.c",
   "native/zero-c/src/emit_coff.c",
