+++ scripts/row-syntax-smoke.mts
@@ -22,7 +22,7 @@ try {
     "-Wpedantic",
     "-I",
     "native/zero-c/include",
-    "native/zero-c/src/parser.c",
+    "native/zero-c/src/ast.c",
     "native/zero-c/src/row_syntax.c",
     "native/zero-c/tests/row_syntax_smoke.c",
     "-o",
@@ -41,7 +41,7 @@ try {
     "-I",
     "native/zero-c/include",
     "native/zero-c/src/fs.c",
-    "native/zero-c/src/parser.c",
+    "native/zero-c/src/ast.c",
     "native/zero-c/src/row_syntax.c",
     "native/zero-c/src/checker.c",
     "native/zero-c/src/type_core.c",
